client: use proxy to determine default for http2
The current implementation of HTTP/2 does not handle proxy servers. Use `HttpClient.findProxyFromEnvironment` to see if a request to Figma would be proxied.

Refactor the `FigmaClient` so the `_send` method is now a function pointer. Use conditional imports to handle the lack of HTTP/2 in the browser by exposing a getter for a `http2` implementation. On browser it always returns `null` but in `io` if a proxy is not used it returns the HTTP2 implementation.